addssaving throw display and refactors more
